Move logic out of `cli.rs` into a separate module `commands.rs`.
