+++ test-utils/state-viewer/src/lib.rs
@@ -1,6 +1,7 @@
 #![doc = include_str!("../README.md")]
 
 mod apply_chain_range;
+mod commands;
 mod state_dump;
 
 pub mod cli;
