+++ test-utils/state-viewer/src/main.rs
@@ -1,5 +1,6 @@
 mod apply_chain_range;
 mod cli;
+mod commands;
 mod state_dump;
 
 fn main() {
